+++ CommonConcepts/Plugins/Rhetos.Dsl.DefaultConcepts/Computations/ItemFilterApplyOnClientReadInfo.cs
@@ -36,7 +36,7 @@ namespace Rhetos.Dsl.DefaultConcepts
         [ConceptKey]
         public ItemFilterInfo Filter { get; set; }
 
-        public IEnumerable<IConceptInfo> CreateNewConcepts(IEnumerable<IConceptInfo> existingConcepts)
+        public IEnumerable<IConceptInfo> CreateNewConcepts()
         {
             return new IConceptInfo[]
             {