+++ CommonConcepts/Plugins/Rhetos.Dsl.DefaultConcepts/SimpleBusinessLogic/MinValueInfo.cs
@@ -48,7 +48,7 @@ namespace Rhetos.Dsl.DefaultConcepts
             { typeof(DateTimePropertyInfo), limit => String.Format(@"var limit = DateTime.Parse({0})", CsUtility.QuotedString(limit)) },
         };
 
-        public IEnumerable<IConceptInfo> CreateNewConcepts(IEnumerable<IConceptInfo> existingConcepts)
+        public IEnumerable<IConceptInfo> CreateNewConcepts()
         {
             string limitSnippet = LimitSnippetByType
                 .Where(snippet => snippet.Key.IsAssignableFrom(Property.GetType()))